+++ packages/beacon-node/test/e2e/network/peers/peerManager.test.ts
@@ -11,7 +11,7 @@ import {PeerRpcScoreStore, PeerManager, IReqRespBeaconNodePeerManager} from "../
 import {Eth2Gossipsub, getConnectionsMap, NetworkEvent, NetworkEventBus} from "../../../../src/network/index.js";
 import {PeersData} from "../../../../src/network/peers/peersData.js";
 import {createNode} from "../../../utils/network.js";
-import {getAttnets, getSyncnets} from "../../../utils/networkData.js";
+import {getAttnets, getSyncnets} from "../../../utils/network.js";
 import {generateState} from "../../../utils/state.js";
 import {waitForEvent} from "../../../utils/events/resolver.js";
 import {testLogger} from "../../../utils/logger.js";
