+++ packages/beacon-node/test/e2e/network/reqresp.test.ts
@@ -7,7 +7,8 @@ import {allForks, altair, phase0, Root, ssz} from "@lodestar/types";
 import {sleep as _sleep} from "@lodestar/utils";
 import {Network, ReqRespBeaconNodeOpts} from "../../../src/network/index.js";
 import {expectRejectedWithLodestarError} from "../../utils/errors.js";
-import {connect, getNetworkForTest, getPeerIdOf, onPeerConnect} from "../../utils/network.js";
+import {connect, getPeerIdOf, onPeerConnect} from "../../utils/network.js";
+import {getNetworkForTest} from "../../utils/networkWithMockDb.js";
 import {arrToSource} from "../../unit/network/reqresp/utils.js";
 import {GetReqRespHandlerFn, ReqRespMethod} from "../../../src/network/reqresp/types.js";
 import {PeerIdStr} from "../../../src/util/peerId.js";
